requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
psycopg2==2.8.4
python-dotenv==0.17.0
Werkzeug==0.16.0